+++ app/kumactl/cmd/get/get_healthchecks_test.go
@@ -90,6 +90,7 @@ var _ = Describe("kumactl get healthchecks", func() {
 		type testCase struct {
 			outputFormat string
 			goldenFile   string
+			pagination   string
 			matcher      func(interface{}) gomega_types.GomegaMatcher
 		}
 
@@ -98,7 +99,7 @@ var _ = Describe("kumactl get healthchecks", func() {
 				// given
 				rootCmd.SetArgs(append([]string{
 					"--config-file", filepath.Join("..", "testdata", "sample-kumactl.config.yaml"),
-					"get", "healthchecks"}, given.outputFormat))
+					"get", "healthchecks"}, given.outputFormat, given.pagination))
 
 				// when
 				err := rootCmd.Execute()
@@ -122,6 +123,14 @@ var _ = Describe("kumactl get healthchecks", func() {
 			Entry("should support Table output explicitly", testCase{
 				outputFormat: "-otable",
 				goldenFile:   "get-healthchecks.golden.txt",
+				matcher: func(expected interface{}) gomega_types.GomegaMatcher {
+					return WithTransform(strings.TrimSpace, Equal(strings.TrimSpace(string(expected.([]byte)))))
+				},
+			}),
+			Entry("should support pagination", testCase{
+				outputFormat: "-otable",
+				pagination:   "--size=1",
+				goldenFile:   "get-healthchecks.pagination.golden.txt",
 				matcher: func(expected interface{}) gomega_types.GomegaMatcher {
 					return WithTransform(strings.TrimSpace, Equal(strings.TrimSpace(string(expected.([]byte)))))
 				},