+++ app/kumactl/cmd/get/get_traffic_route.go
@@ -35,7 +35,9 @@ func newGetTrafficRouteCmd(pctx *getContext) *cobra.Command {
 				}
 				return errors.Wrapf(err, "failed to get mesh %s", currentMesh)
 			}
-			trafficRoutes := []*mesh.TrafficRouteResource{trafficRoute}
+			trafficRoutes := &mesh.TrafficRouteResourceList{
+				Items: []*mesh.TrafficRouteResource{trafficRoute},
+			}
 			switch format := output.Format(pctx.args.outputFormat); format {
 			case output.TableFormat:
 				return printTrafficRoutes(trafficRoutes, cmd.OutOrStdout())