+++ pkg/plugins/resources/remote/unmarshalling.go
@@ -2,7 +2,10 @@ package remote
 
 import (
 	"encoding/json"
+	"net/url"
 	"time"
+
+	"github.com/pkg/errors"
 
 	"github.com/Kong/kuma/pkg/core/resources/model"
 	"github.com/Kong/kuma/pkg/core/resources/model/rest"
@@ -72,6 +75,20 @@ func UnmarshalList(b []byte, rs model.ResourceList) error {
 			ModificationTime: ri.Meta.ModificationTime,
 		})
 		_ = rs.AddItem(r)
+	}
+	if rsr.Next != nil {
+		uri, err := url.ParseRequestURI(*rsr.Next)
+		if err != nil {
+			return errors.Wrap(err, "invalid next URL from the server")
+		}
+		offset := uri.Query().Get("offset")
+		// we do not preserve here the size of the page, but since it is used in kumactl
+		// user will rerun command with the page size of his choice
+		if offset != "" {
+			rs.SetPagination(model.Pagination{
+				NextOffset: offset,
+			})
+		}
 	}
 	return nil
 }