+++ pkg/plugins/secrets/k8s/store_test.go
@@ -14,6 +14,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	system_proto "github.com/Kong/kuma/api/system/v1alpha1"
 	core_system "github.com/Kong/kuma/pkg/core/resources/apis/system"
 	secret_model "github.com/Kong/kuma/pkg/core/resources/apis/system"
 	"github.com/Kong/kuma/pkg/core/resources/store"
@@ -82,26 +83,31 @@ var _ = Describe("KubernetesStore", func() {
 		It("should create a new secret", func() {
 			// given
 			secret := &secret_model.SecretResource{
-				Spec: wrappers.BytesValue{
-					Value: []byte("example"),
+				Spec: system_proto.Secret{
+					Data: &wrappers.BytesValue{
+						Value: []byte("example"),
+					},
 				},
 			}
 			expected := backend.ParseYAML(`
             apiVersion: v1
             kind: Secret
             type: system.kuma.io/secret
+            metadata:
+              labels:
+                kuma.io/mesh: demo
             data:
               value: ZXhhbXBsZQ== # base64(example)
 `).(*kube_core.Secret)
 
 			// when
-			err := s.Create(context.Background(), secret, store.CreateByKey(name, noMesh))
+			err := s.Create(context.Background(), secret, store.CreateByKey(name, "demo"))
 
 			// then
 			Expect(err).ToNot(HaveOccurred())
 			// and
 			Expect(secret.Meta.GetName()).To(Equal(name))
-			Expect(secret.Meta.GetMesh()).To(Equal(""))
+			Expect(secret.Meta.GetMesh()).To(Equal("demo"))
 			Expect(secret.Meta.GetVersion()).ToNot(Equal(""))
 
 			// when
@@ -143,6 +149,8 @@ var _ = Describe("KubernetesStore", func() {
             metadata:
               namespace: %s
               name: %s
+              labels:
+                kuma.io/mesh: demo
             data:
               value: ZXhhbXBsZQ== # base64(example)
 `, ns, name))
@@ -160,13 +168,13 @@ var _ = Describe("KubernetesStore", func() {
 			secret := &secret_model.SecretResource{}
 
 			// when
-			err := s.Get(context.Background(), secret, store.GetByKey(name, noMesh))
+			err := s.Get(context.Background(), secret, store.GetByKey(name, "demo"))
 			// then
 			Expect(err).ToNot(HaveOccurred())
 			version := secret.Meta.GetVersion()
 
 			// when
-			secret.Spec.Value = []byte("another")
+			secret.Spec.Data.Value = []byte("another")
 			err = s.Update(context.Background(), secret)
 
 			// then
@@ -195,6 +203,8 @@ var _ = Describe("KubernetesStore", func() {
             metadata:
               namespace: %s
               name: %s
+              labels:
+                kuma.io/mesh: demo
 `, ns, name))
 			backend.Create(initial)
 
@@ -202,7 +212,7 @@ var _ = Describe("KubernetesStore", func() {
 			secret := &secret_model.SecretResource{}
 
 			// when
-			err := s.Get(context.Background(), secret, store.GetByKey(name, noMesh))
+			err := s.Get(context.Background(), secret, store.GetByKey(name, "demo"))
 			// then
 			Expect(err).ToNot(HaveOccurred())
 
@@ -212,7 +222,7 @@ var _ = Describe("KubernetesStore", func() {
 			err = s.Update(context.Background(), secret)
 
 			// then
-			Expect(err).To(MatchError(store.ErrorResourceConflict(core_system.SecretType, name, noMesh)))
+			Expect(err).To(MatchError(store.ErrorResourceConflict(core_system.SecretType, name, "demo")))
 		})
 
 		It("should return an error if resource has changed", func() {
@@ -224,6 +234,8 @@ var _ = Describe("KubernetesStore", func() {
             metadata:
               namespace: %s
               name: %s
+              labels:
+                kuma.io/mesh: demo             
 `, ns, name))
 			backend.Create(initial)
 
@@ -231,7 +243,7 @@ var _ = Describe("KubernetesStore", func() {
 			secret1 := &secret_model.SecretResource{}
 
 			// when
-			err := s.Get(context.Background(), secret1, store.GetByKey(name, noMesh))
+			err := s.Get(context.Background(), secret1, store.GetByKey(name, "demo"))
 			// then
 			Expect(err).ToNot(HaveOccurred())
 
@@ -239,21 +251,29 @@ var _ = Describe("KubernetesStore", func() {
 			secret2 := &secret_model.SecretResource{}
 
 			// when
-			err = s.Get(context.Background(), secret2, store.GetByKey(name, noMesh))
+			err = s.Get(context.Background(), secret2, store.GetByKey(name, "demo"))
 			// then
 			Expect(err).ToNot(HaveOccurred())
 
 			// when
-			secret1.Spec.Value = []byte("example")
+			secret1.Spec = system_proto.Secret{
+				Data: &wrappers.BytesValue{
+					Value: []byte("example"),
+				},
+			}
 			err = s.Update(context.Background(), secret1)
 			// then
 			Expect(err).ToNot(HaveOccurred())
 
 			// when
-			secret2.Spec.Value = []byte("another")
+			secret2.Spec = system_proto.Secret{
+				Data: &wrappers.BytesValue{
+					Value: []byte("another"),
+				},
+			}
 			err = s.Update(context.Background(), secret2)
 			// then
-			Expect(err).To(MatchError(store.ErrorResourceConflict(core_system.SecretType, name, noMesh)))
+			Expect(err).To(MatchError(store.ErrorResourceConflict(core_system.SecretType, name, "demo")))
 		})
 	})
 
@@ -263,10 +283,10 @@ var _ = Describe("KubernetesStore", func() {
 			backend.AssertNotExists(&kube_core.Secret{}, ns, name)
 
 			// when
-			err := s.Get(context.Background(), &secret_model.SecretResource{}, store.GetByKey(name, noMesh))
+			err := s.Get(context.Background(), &secret_model.SecretResource{}, store.GetByKey(name, "demo"))
 
 			// then
-			Expect(err).To(MatchError(store.ErrorResourceNotFound(core_system.SecretType, name, noMesh)))
+			Expect(err).To(MatchError(store.ErrorResourceNotFound(core_system.SecretType, name, "demo")))
 		})
 
 		It("should return an existing resource", func() {
@@ -278,6 +298,8 @@ var _ = Describe("KubernetesStore", func() {
             metadata:
               namespace: %s
               name: %s
+              labels:
+                kuma.io/mesh: demo
             data:
               value: ZXhhbXBsZQ== # base64(example)
 `, ns, name))
@@ -287,14 +309,15 @@ var _ = Describe("KubernetesStore", func() {
 			actual := &secret_model.SecretResource{}
 
 			// when
-			err := s.Get(context.Background(), actual, store.GetByKey(name, noMesh))
+			err := s.Get(context.Background(), actual, store.GetByKey(name, "demo"))
 
 			// then
 			Expect(err).ToNot(HaveOccurred())
 			// and
 			Expect(actual.Meta.GetName()).To(Equal(name))
+			Expect(actual.Meta.GetMesh()).To(Equal("demo"))
 			// and
-			Expect(actual.Spec.Value).To(Equal([]byte("example")))
+			Expect(actual.Spec.Data.Value).To(Equal([]byte("example")))
 		})
 	})
 
@@ -304,7 +327,7 @@ var _ = Describe("KubernetesStore", func() {
 			backend.AssertNotExists(&kube_core.Secret{}, ns, name)
 
 			// when
-			err := s.Delete(context.Background(), &secret_model.SecretResource{}, store.DeleteByKey(name, noMesh))
+			err := s.Delete(context.Background(), &secret_model.SecretResource{}, store.DeleteByKey(name, "demo"))
 
 			// then
 			Expect(err).ToNot(HaveOccurred())
@@ -319,11 +342,13 @@ var _ = Describe("KubernetesStore", func() {
             metadata:
               namespace: %s
               name: %s
+              labels:
+                kuma.io/mesh: demo
 `, ns, name))
 			backend.Create(initial)
 
 			// when
-			err := s.Delete(context.Background(), &secret_model.SecretResource{}, store.DeleteByKey(name, noMesh))
+			err := s.Delete(context.Background(), &secret_model.SecretResource{}, store.DeleteByKey(name, "demo"))
 
 			// then
 			Expect(err).ToNot(HaveOccurred())
@@ -346,52 +371,88 @@ var _ = Describe("KubernetesStore", func() {
 			Expect(secrets.Items).To(HaveLen(0))
 		})
 
-		It("should return a list of matching resource", func() {
-			// setup
-			one := backend.ParseYAML(fmt.Sprintf(`
-            apiVersion: v1
-            kind: Secret
-            type: system.kuma.io/secret
-            metadata:
-              namespace: %s
-              name: %s
-            data:
-              value: ZXhhbXBsZQ== # base64(example)
+		Describe("with resources loaded", func() {
+			BeforeEach(func() {
+				// given secret in demo mesh
+				one := backend.ParseYAML(fmt.Sprintf(`
+                apiVersion: v1
+                kind: Secret
+                type: system.kuma.io/secret
+                metadata:
+                  namespace: %s
+                  name: %s
+                  labels:
+                    kuma.io/mesh: demo
+                data:
+                  value: ZXhhbXBsZQ== # base64(example)
 `, ns, "one"))
-			backend.Create(one)
-			// and
-			two := backend.ParseYAML(fmt.Sprintf(`
-            apiVersion: v1
-            kind: Secret
-            type: system.kuma.io/secret
-            metadata:
-              namespace: %s
-              name: %s
-            data:
-              value: YW5vdGhlcg== # base64(another)
+				backend.Create(one)
+				// and secret in default mesh
+				two := backend.ParseYAML(fmt.Sprintf(`
+                apiVersion: v1
+                kind: Secret
+                type: system.kuma.io/secret
+                metadata:
+                  namespace: %s
+                  name: %s
+                  labels:
+                    kuma.io/mesh: default
+                data:
+                  value: YW5vdGhlcg== # base64(another)
 `, ns, "two"))
-			backend.Create(two)
+				backend.Create(two)
+				// and secret which is not a Kuma secret but resides in the kuma-system namespace - this should be ignored
+				three := backend.ParseYAML(fmt.Sprintf(`
+                apiVersion: v1
+                kind: Secret
+                type: some-other-type
+                metadata:
+                  namespace: %s
+                  name: %s
+                  labels:
+                    kuma.io/mesh: default
+                data:
+                  value: YW5vdGhlcg== # base64(another)
+`, ns, "three"))
+				backend.Create(three)
+			})
 
-			// given
-			secrets := &secret_model.SecretResourceList{}
+			It("should return a list of secrets in all meshes", func() {
+				// given
+				secrets := &secret_model.SecretResourceList{}
 
-			// when
-			err := s.List(context.Background(), secrets)
+				// when
+				err := s.List(context.Background(), secrets)
 
-			// then
-			Expect(err).ToNot(HaveOccurred())
-			// and
-			Expect(secrets.Items).To(HaveLen(2))
+				// then
+				Expect(err).ToNot(HaveOccurred())
+				// and
+				Expect(secrets.Items).To(HaveLen(2))
 
-			// when
-			items := map[string]*secret_model.SecretResource{
-				secrets.Items[0].Meta.GetName(): secrets.Items[0],
-				secrets.Items[1].Meta.GetName(): secrets.Items[1],
-			}
-			// then
-			Expect(items["one"].Spec.Value).To(Equal([]byte("example")))
-			// and
-			Expect(items["two"].Spec.Value).To(Equal([]byte("another")))
+				// when
+				items := map[string]*secret_model.SecretResource{
+					secrets.Items[0].Meta.GetName(): secrets.Items[0],
+					secrets.Items[1].Meta.GetName(): secrets.Items[1],
+				}
+				// then
+				Expect(items["one"].Spec.Data.Value).To(Equal([]byte("example")))
+				// and
+				Expect(items["two"].Spec.Data.Value).To(Equal([]byte("another")))
+			})
+
+			It("should return a list of secrets in given mesh", func() {
+				// given
+				secrets := &secret_model.SecretResourceList{}
+
+				// when
+				err := s.List(context.Background(), secrets, store.ListByMesh("default"))
+
+				// then
+				Expect(err).ToNot(HaveOccurred())
+				Expect(secrets.Items).To(HaveLen(1))
+				Expect(secrets.Items[0].Spec.Data.Value).To(Equal([]byte("another")))
+			})
 		})
+
 	})
 })